Update bindings to account for removed Path class (#1217)
