--- Gui/opensim/modeling/src/org/opensim/modeling/Path.java
@@ -1,70 +1,0 @@
-/* ----------------------------------------------------------------------------
- * This file was automatically generated by SWIG (http://www.swig.org).
- * Version 3.0.9
- *
- * Do not make changes to this file unless you know what you are doing--modify
- * the SWIG interface file instead.
- * ----------------------------------------------------------------------------- */
-
-package org.opensim.modeling;
-
-public class Path {
-  private transient long swigCPtr;
-  protected transient boolean swigCMemOwn;
-
-  public Path(long cPtr, boolean cMemoryOwn) {
-    swigCMemOwn = cMemoryOwn;
-    swigCPtr = cPtr;
-  }
-
-  public static long getCPtr(Path obj) {
-    return (obj == null) ? 0 : obj.swigCPtr;
-  }
-
-  protected void finalize() {
-    delete();
-  }
-
-  public synchronized void delete() {
-    if (swigCPtr != 0) {
-      if (swigCMemOwn) {
-        swigCMemOwn = false;
-        opensimCommonJNI.delete_Path(swigCPtr);
-      }
-      swigCPtr = 0;
-    }
-  }
-
-  public String toString() {
-    return opensimCommonJNI.Path_toString(swigCPtr, this);
-  }
-
-  public boolean isAbsolute() {
-    return opensimCommonJNI.Path_isAbsolute(swigCPtr, this);
-  }
-
-  public long getNumPathLevels() {
-    return opensimCommonJNI.Path_getNumPathLevels(swigCPtr, this);
-  }
-
-  public void pushBack(String pathElement) {
-    opensimCommonJNI.Path_pushBack(swigCPtr, this, pathElement);
-  }
-
-  public char getSeparator() {
-    return opensimCommonJNI.Path_getSeparator(swigCPtr, this);
-  }
-
-  public String getInvalidChars() {
-    return opensimCommonJNI.Path_getInvalidChars(swigCPtr, this);
-  }
-
-  public boolean isLegalPathElement(String pathElement) {
-    return opensimCommonJNI.Path_isLegalPathElement(swigCPtr, this, pathElement);
-  }
-
-  public void trimDotAndDotDotElements() {
-    opensimCommonJNI.Path_trimDotAndDotDotElements(swigCPtr, this);
-  }
-
-}
