--- Gui/opensim/modeling/src/org/opensim/modeling/SWIGTYPE_p_std__shared_ptrT_OpenSim__LogSink_t.java
@@ -1,26 +1,0 @@
-/* ----------------------------------------------------------------------------
- * This file was automatically generated by SWIG (http://www.swig.org).
- * Version 3.0.9
- *
- * Do not make changes to this file unless you know what you are doing--modify
- * the SWIG interface file instead.
- * ----------------------------------------------------------------------------- */
-
-package org.opensim.modeling;
-
-public class SWIGTYPE_p_std__shared_ptrT_OpenSim__LogSink_t {
-  private transient long swigCPtr;
-
-  protected SWIGTYPE_p_std__shared_ptrT_OpenSim__LogSink_t(long cPtr, @SuppressWarnings("unused") boolean futureUse) {
-    swigCPtr = cPtr;
-  }
-
-  protected SWIGTYPE_p_std__shared_ptrT_OpenSim__LogSink_t() {
-    swigCPtr = 0;
-  }
-
-  protected static long getCPtr(SWIGTYPE_p_std__shared_ptrT_OpenSim__LogSink_t obj) {
-    return (obj == null) ? 0 : obj.swigCPtr;
-  }
-}
-
